WS_04/HealthCare_Chatbot/requirements.txt
langchain
langchain-core
openai
pinecone-client
streamlit
python-dotenv
tqdm
